Update MSFT_xADDomain.psm1
Get-TargetResource:
Fixed the Write-Verbose message on line 50: It had the wrong variable for "$dc.ParentDomainName", replaced it with "$domain.ParentDomain"

Test-TargetResource:
Fixed the function always returned false when checking for a child domain with a parent domain. This was happening as Get-TargetResource returns the fully qualified domain name Eg "<childdomainname>.<parent.domain>" and the Test-TargetResource was comparing the fully qualified domain name to the child domain name only.